+++ tomcat/src/main/java/handler/Controller.java
@@ -1,4 +1,7 @@
-package org.apache.coyote.http11;
+package handler;
+
+import org.apache.coyote.http11.HttpRequest;
+import org.apache.coyote.http11.HttpResponse;
 
 public interface Controller {
