+++ tomcat/src/main/java/handler/RequestHandler.java
@@ -1,0 +1,6 @@
+package handler;
+
+public interface RequestHandler {
+
+    Controller getHandler(final String requestUri);
+}
