+++ tomcat/src/main/java/nextstep/jwp/controller/LoginController.java
@@ -1,8 +1,8 @@
 package nextstep.jwp.controller;
 
+import handler.Controller;
 import java.util.Map;
 import nextstep.jwp.db.InMemoryUserRepository;
-import org.apache.coyote.http11.Controller;
 import org.apache.coyote.http11.HttpMethod;
 import org.apache.coyote.http11.HttpRequest;
 import org.apache.coyote.http11.HttpResponse;
