+++ tomcat/src/main/java/nextstep/jwp/controller/RegisterController.java
@@ -1,9 +1,9 @@
 package nextstep.jwp.controller;
 
+import handler.Controller;
 import java.util.Map;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
-import org.apache.coyote.http11.Controller;
 import org.apache.coyote.http11.HttpMethod;
 import org.apache.coyote.http11.HttpRequest;
 import org.apache.coyote.http11.HttpResponse;
